SDK-956: Renamed entities and created documentation
