+++ docs/manifest_validation.md
@@ -1,0 +1,14 @@
+## Manifest validation
+
+Manifest validation is a feature that validate YAML structure definition for task.
+It checks structure, types and related entities.
+
+### How to use
+
+Simply run `spryker-sdk sdk:validate:task` to validate task yaml definition.
+
+### How to fix
+
+ - Task [conventions](conventions.md#Task), [structure](task.md#Task)
+ - Task Set [conventions](conventions.md#Task Set), [structure](task_set.md#Task set)
+
