+++ src/Core/Application/Resources/config/services.yaml
@@ -152,7 +152,7 @@ services:
       - !tagged_iterator report.generator_resolver
 
   manifest_validator_factory:
-    class: SprykerSdk\Sdk\Core\Application\Service\ManifestNormalizer\ManifestNormaliserRegistry
+    class: SprykerSdk\Sdk\Core\Application\Service\ManifestValidator\ManifestValidatorRegistry
     arguments:
       - !tagged_iterator manifest.validator
 
@@ -160,7 +160,7 @@ services:
     class: Symfony\Component\Config\Definition\Processor
 
   manifest_validation:
-    class: SprykerSdk\Sdk\Core\Application\Service\ManifestNormalizer\ManifestNormalizer
+    class: SprykerSdk\Sdk\Core\Application\Service\ManifestValidator\ManifestValidator
     arguments:
       - '@manifest_validator_factory'
       - '@config_processor'
