+++ src/Extension/Resources/config/task/AcpAppAsyncApiCodeGenerateTask.yaml
@@ -3,7 +3,7 @@ id: "acp:asyncapi:generate"
 short_description: "Generate ACP AsyncAPI code."
 help: ~
 stage: build
-version: 0.2.0
+version: 0.2.1
 command: "INSTALLED_ROOT_DIRECTORY=%sdk_dir% php %sdk_dir%/vendor/bin/asyncapi code:asyncapi:generate %asyncapi-file% -v"
 type: local_cli_interactive
 placeholders:
